Module02/airflow/requirements.txt
apache-airflow-providers-google
requests
pyarrow
google-cloud-bigquery
